Moving all paging async stream types to azure-core-rest
