+++ sdk/core/azure-core-rest/build.gradle
@@ -24,6 +24,7 @@ dependencies {
     // <!-- begin: test Dependencies -->
     testImplementation project (":sdk:core:azure-core-jackson")
 
+    testImplementation "net.sourceforge.streamsupport:android-retrofuture:$retroFutureVersion"
     testImplementation "org.junit.jupiter:junit-jupiter-api:$junitJupiterVersion"
     testImplementation "org.junit.jupiter:junit-jupiter-params:$junitJupiterVersion"
     testImplementation "org.mockito:mockito-core:$mockitoVersion"
